Health Facility and Administrative Units Table

A comprehensive list of health facilities and administrative hierarchies, including ownership, operational status, and geographic coordinates.

health_facility_table

Format

A data frame with 8676 rows and 22 columns:

location_name

Name of Health facility

location_id

DHIS2 assigned ID for health facility

parent_id

DHIS2 assigned ID of parent health facility (subcounty)

level

DHIS2 assigned hierarchical level

authority

The governing body or authority responsible for the facility

facility_level

The original classification level of the health facility

medical_bureaus

The medical bureau affiliation (e.g., UPMB, UCMB)

operational_status

Current status of the facility (e.g., Functional, Non-Functional)

ownership

Type of ownership (e.g., Public, Private, PNFP)

private_facilities

Indicator for private sector classification (PFP, PNFP)

public_facilities

Indicator for public sector classification (e.g. MOH,BOU, Local Government)

reporting_status

Current status of data reporting compliance

path_to_top_parent

String of DHIS2 IDs from current health facility to the top-level parent

region_name

Name of region associated with health facility

district_name

Name of district associated with health facility

DLG_name

Name of District Local Government (DLG) associated with health facility

subcounty_name

Name of subcounty associated with health facility

latitude

Geographic coordinate: North-South position

longitude

Geographic coordinate: East-West position

geom_enhanced

Indicator if geometry was added through external web-scraping and not present in DHIS2

facility_level_2

Secondary facility level classification (clinic, hospital, drugshop, other)

facility_type

Secondary facility label created to match other surveys on health seeking (private, gov_HC, private_hosp, drugshop, gov_hosp, other)